refactor(slack): address review — drop dead assistant-scope code, extract tool summaries
- Remove assistant_thread event registrations + handlers (only fire with
  assistant:write, which we no longer request)
- Remove unused isModernSlackAiAgentEnabled + its featureFlagModel dependency
- Document getCardImage as intentionally public (Slack unfurl); safeguards
  (nanoid + no-store/noindex headers) were already in place
- Extract tool-summary helpers to ai/utils/toolSummaries.ts
