+++ packages/backend/src/ee/services/AiAgentService/AiAgentService.ts
@@ -8394,79 +8394,6 @@ Use your existing tools to inspect them when relevant to the user's question. Wh
         );
     }
 
-    public async handleAssistantThreadStarted({
-        event,
-        context,
-    }: {
-        event: AnyType;
-        context: AnyType;
-    }) {
-        const assistantThread = event?.assistant_thread;
-        const channelId = assistantThread?.channel_id;
-        const threadTs = assistantThread?.thread_ts;
-        const teamId =
-            context?.teamId ??
-            context?.team_id ??
-            assistantThread?.context?.team_id;
-
-        if (!teamId || !channelId || !threadTs) {
-            Logger.debug(
-                'Skipping assistant_thread_started without thread ids',
-                {
-                    teamId,
-                    channelId,
-                    threadTs,
-                },
-            );
-            return;
-        }
-
-        try {
-            const organizationUuid =
-                await this.slackAuthenticationModel.getOrganizationUuidFromTeamId(
-                    teamId,
-                );
-            await this.slackClient.setSuggestedPrompts({
-                organizationUuid,
-                channelId,
-                threadTs,
-                title: 'Ask Lightdash',
-                prompts: [
-                    {
-                        title: 'Find the right metric',
-                        message: 'Which metric should I use for this question?',
-                    },
-                    {
-                        title: 'Explain this dashboard',
-                        message:
-                            'Explain the main takeaways from this dashboard.',
-                    },
-                    {
-                        title: 'Build a chart',
-                        message: 'Build a chart that answers this question.',
-                    },
-                    {
-                        title: 'Improve project context',
-                        message: 'What context is missing from this project?',
-                    },
-                ],
-            });
-        } catch (error) {
-            Logger.error('Failed to handle assistant_thread_started', error);
-        }
-    }
-
-    public async handleAssistantThreadContextChanged({
-        event,
-    }: {
-        event: AnyType;
-    }) {
-        this.logger.debug('Slack assistant thread context changed', {
-            channelId: event?.assistant_thread?.context?.channel_id,
-            threadTs: event?.assistant_thread?.thread_ts,
-        });
-    }
-
     // eslint-disable-next-line class-methods-use-this
     public handleClickOAuthButton(app: App) {
         // Match action_id pattern: actions.oauth_button_click:teamId:channelId:messageTs
